Add fields for customizing import and extra response data
Added post type, status, and author fields for importing. Also
added elements to response data for displaying information about
successfully imported posts. This includes permalinks and edit
links for each of the imported posts, as well as the type, status,
and author of all the posts combined.

The output of the type, status, and author is redundant,
because of the possiblity that an invalid post type, status, or
author could be sent in the POST request. It mitigates any
potential confusion when such a situtation occurs.